+++ src/main/java/org/vicky/starterkits/client/ClientConfigHolder.java
@@ -1,0 +1,11 @@
+package org.vicky.starterkits.client;
+
+import java.util.List;
+
+public class ClientConfigHolder {
+    public static String kitSelectorItemName;
+    public static List<String> kitSelectorItemLore;
+    public static int kitMaxUsages;
+    public static boolean kitIsSelectable;
+    public static boolean allowRollableKits;
+}