+++ src/main/java/org/vicky/starterkits/network/packets/OpenKitSelectorScreenPacket.java
@@ -2,7 +2,10 @@ package org.vicky.starterkits.network.packets;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.network.NetworkEvent;
+import org.vicky.starterkits.client.ClientPacketHandlers;
 import org.vicky.starterkits.client.gui.KitSelectionScreen;
 
 import java.util.function.Supplier;
@@ -17,7 +20,8 @@ public class OpenKitSelectorScreenPacket {
     public static void handle(OpenKitSelectorScreenPacket pkt, Supplier<NetworkEvent.Context> ctx) {
         ctx.get().enqueueWork(() -> {
             if (Minecraft.getInstance().player != null) {
-                Minecraft.getInstance().setScreen(new KitSelectionScreen());
+                DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> ClientPacketHandlers::openKitSelectionScreen);
+
             }
         });
         ctx.get().setPacketHandled(true);